chore: adds manual deployment script
This adds a new script that packages up and deploys the site to netlify.
Due to the way netlify and gitlab oauth applications work we need
predictable names for the site to add as valid redirect uris in gitlab.
However, the preview builds do not give predictable names and gitlab api
does not support programatically adding uris to the redirect uri list.

As a crappy work around I have created 10 sites on netlify with fixed
names that we can use for preview builds. The build script added knows
about these sites and will pick one to deploy to in a way that does not
interfere with other current previews.

It does this by adding metadata to each site which can claim a site for
12 at which point that pipeline is the only one that can deploy to
a site through this script. If all the sites are claimed then it picks
the one with the oldest claim and reclaims it.