feat: integrates the PrivateRoute component
Now when the user is not authenticated / and /settings will redirect to /login prompting the user to login via gitlab. Once logged in it will redirect them back to /.

As part of this I removed the login buttons as they do not do anything as you are either on the login page or logged in.